Abstract parsing of userInput to Parse class
